Updated the app so it works, but no facebook login
So the Facebook login no longer works, but I fixed the previous
problems. The SDK and packages are all installed now, so it works fine!
Will get around to doing the Facebook login properly soon, should not
take too much effort. Just need to add the sign on authorisation with
firebase. Very simple instructions online which show how to do it